pyproject.toml
[project]
name = "barcode-transformer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "pillow>=12.1.1",
    "pylibdmtx>=0.1.10",
    "pystray>=0.19.5",
    "python-barcode[images]>=0.16.1",
    "pyzbar>=0.1.9",
    "reportlab>=4.4.10",
    "setuptools>=82.0.1",
    "zxing-cpp>=3.0.0",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
]
